% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_params.R
\name{mod_params_server}
\alias{mod_params_server}
\title{Module paramètres — Server}
\usage{
mod_params_server(id)
}
\arguments{
\item{id}{identifiant du module}
}
\value{
une \code{reactive} list contenant les paramètres normalisés et traitements
}
\description{
Module paramètres — Server
}
\keyword{internal}
