% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{simuler_puissance}
\alias{simuler_puissance}
\title{Simulation de puissance pour un test t (Welch ou pooled)}
\usage{
simuler_puissance(
  n_rep,
  n_sims,
  mu,
  sd,
  effet,
  alpha,
  alternative = "two.sided",
  equal_var = FALSE,
  effect_mode = "add"
)
}
\arguments{
\item{n_rep}{Nombre de répétitions par traitement}

\item{n_sims}{Nombre d'itérations de simulation}

\item{mu}{Rendement moyen (t/ha) du témoin}

\item{sd}{Écart-type (t/ha)}

\item{effet}{Effet attendu (additif ou multiplicatif)}

\item{alpha}{Niveau de signification (ex.: 0.05)}

\item{alternative}{"two.sided", "greater" ou "less"}

\item{equal_var}{TRUE pour test avec variances égales (pooled),
FALSE pour test de Welch}

\item{effect_mode}{"add" (additif, t/ha) ou "mult" (multiplicatif, proportion)}
}
\value{
Proportion de rejets de H0 = puissance estimée (entre 0 et 1)
}
\description{
Cette fonction effectue une simulation Monte-Carlo de la puissance
pour un test t comparant deux traitements.
}
